Fix Issue #5, 2013-12-08 "Binky" compatibilty issue (un-saveable config)
All onoff configuration switches do not get saved when the user changes
the value. The reason was that I used true|false for "onoff" switches
although only 1|0 was allowed [1] (never noticed this).

It worked for years but the way the configuration gets parsed by DokuWiki
was changed [2] [3] after Binky RC2, so the bug affects the current
DokuWiki Binky final only.

[1] <https://www.dokuwiki.org/config#configuration_options>
[2] <https://github.com/splitbrain/dokuwiki/commit/43e700149db19863b559e0eebc44132c611ef08d>
[3] <http://www.freelists.org/post/dokuwiki/SemiUrgent-broken-template-config-in-final-Binky,1>